feat: refactor splitting Utils.kt into common resources
