+++ state/BUILD
@@ -22,7 +22,7 @@ package(default_visibility = [
     "//module:__subpackages__",
     "//framework:__subpackages__",
     "//:__pkg__",
-    "//test/integration:__pkg__",
+    "//test/integration:__subpackages__",
 ])
 
 kt_jvm_library(
