+++ test/integration/TextEditorTest.kt
@@ -26,25 +26,27 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.vaticle.typedb.client.api.TypeDBSession
 import com.vaticle.typedb.studio.state.StudioState
-import com.vaticle.typedb.studio.test.integration.Utils.studioTest
-import com.vaticle.typedb.studio.test.integration.Utils.studioTestWithRunner
-import com.vaticle.typedb.studio.test.integration.Utils.connectToTypeDB
-import com.vaticle.typedb.studio.test.integration.Utils.createDatabase
-import com.vaticle.typedb.studio.test.integration.Utils.cloneAndOpenProject
-import com.vaticle.typedb.studio.test.integration.Utils.delayAndRecompose
-import com.vaticle.typedb.studio.test.integration.Utils.writeSchemaInteractively
-import com.vaticle.typedb.studio.test.integration.Utils.writeDataInteractively
-import com.vaticle.typedb.studio.test.integration.Utils.verifyDataWrite
-import com.vaticle.typedb.studio.test.integration.Utils.SCHEMA_FILE_NAME
-import com.vaticle.typedb.studio.test.integration.Utils.DATA_FILE_NAME
-import com.vaticle.typedb.studio.test.integration.Utils.QUERY_FILE_NAME
-import com.vaticle.typedb.studio.test.integration.Utils.TQL_DATA_PATH
-import com.vaticle.typedb.studio.test.integration.Utils.SAVE_ICON_STRING
-import com.vaticle.typedb.studio.test.integration.Utils.PLAY_ICON_STRING
-import com.vaticle.typedb.studio.test.integration.Utils.ROLLBACK_ICON_STRING
-import com.vaticle.typedb.studio.test.integration.Utils.PLUS_ICON_STRING
-import com.vaticle.typedb.studio.test.integration.Utils.SAMPLE_DATA_PATH
-import com.vaticle.typedb.studio.test.integration.Utils.CHEVRON_UP_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.CHEVRON_UP_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.DATA_FILE_NAME
+import com.vaticle.typedb.studio.test.integration.common.Data.PLAY_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.PLUS_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.QUERY_FILE_NAME
+import com.vaticle.typedb.studio.test.integration.common.Data.ROLLBACK_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.SAMPLE_DATA_PATH
+import com.vaticle.typedb.studio.test.integration.common.Data.SAVE_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.SCHEMA_FILE_NAME
+import com.vaticle.typedb.studio.test.integration.common.Data.TQL_DATA_PATH
+import com.vaticle.typedb.studio.test.integration.common.Delays.FILE_IO
+import com.vaticle.typedb.studio.test.integration.common.Delays.NETWORK_IO
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.cloneAndOpenProject
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.connectToTypeDB
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.createDatabase
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.delayAndRecompose
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.verifyDataWrite
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.writeDataInteractively
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.writeSchemaInteractively
+import com.vaticle.typedb.studio.test.integration.common.StudioTestHelpers.studioTest
+import com.vaticle.typedb.studio.test.integration.common.StudioTestHelpers.studioTestWithRunner
 import java.io.File
 import kotlin.test.assertTrue
 import org.junit.Test
@@ -65,7 +67,7 @@ class TextEditorTest: IntegrationTest() {
             val file = File("$path/Untitled1.tql")
             StudioState.project.saveFileDialog.file!!.trySave(file.toPath(), true)
             StudioState.project.current!!.reloadEntries()
-            delayAndRecompose(composeRule, Delays.FILE_IO)
+            delayAndRecompose(composeRule, FILE_IO)
 
             assertTrue(file.exists())
         }
@@ -81,7 +83,7 @@ class TextEditorTest: IntegrationTest() {
             writeSchemaInteractively(composeRule, dbName = testID, SCHEMA_FILE_NAME)
 
             StudioState.client.session.tryOpen(database = testID, TypeDBSession.Type.DATA)
-            delayAndRecompose(composeRule, Delays.NETWORK_IO)
+            delayAndRecompose(composeRule, NETWORK_IO)
 
             composeRule.onNodeWithText(CHEVRON_UP_ICON_STRING).performClick()
             delayAndRecompose(composeRule)
@@ -112,7 +114,7 @@ class TextEditorTest: IntegrationTest() {
             createDatabase(composeRule, dbName = testID)
 
             StudioState.client.session.tryOpen(testID, TypeDBSession.Type.SCHEMA)
-            delayAndRecompose(composeRule, Delays.NETWORK_IO)
+            delayAndRecompose(composeRule, NETWORK_IO)
 
             composeRule.onNodeWithText("schema").performClick()
             composeRule.onNodeWithText("write").performClick()
@@ -120,9 +122,9 @@ class TextEditorTest: IntegrationTest() {
             StudioState.project.current!!.directory.entries.find { it.name == SCHEMA_FILE_NAME }!!.asFile().tryOpen()
 
             composeRule.onNodeWithText(PLAY_ICON_STRING).performClick()
-            delayAndRecompose(composeRule, Delays.NETWORK_IO)
+            delayAndRecompose(composeRule, NETWORK_IO)
             composeRule.onNodeWithText(ROLLBACK_ICON_STRING).performClick()
-            delayAndRecompose(composeRule, Delays.NETWORK_IO)
+            delayAndRecompose(composeRule, NETWORK_IO)
 
             composeRule.onNodeWithText("repo-id").assertDoesNotExist()
         }
