+++ test/integration/TypeBrowserTest.kt
@@ -27,16 +27,17 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.vaticle.typedb.client.api.TypeDBSession
 import com.vaticle.typedb.studio.state.StudioState
-import com.vaticle.typedb.studio.test.integration.Utils.studioTestWithRunner
-import com.vaticle.typedb.studio.test.integration.Utils.connectToTypeDB
-import com.vaticle.typedb.studio.test.integration.Utils.createDatabase
-import com.vaticle.typedb.studio.test.integration.Utils.cloneAndOpenProject
-import com.vaticle.typedb.studio.test.integration.Utils.delayAndRecompose
-import com.vaticle.typedb.studio.test.integration.Utils.writeSchemaInteractively
-import com.vaticle.typedb.studio.test.integration.Utils.SCHEMA_FILE_NAME
-import com.vaticle.typedb.studio.test.integration.Utils.TQL_DATA_PATH
-import com.vaticle.typedb.studio.test.integration.Utils.DOUBLE_CHEVRON_DOWN_ICON_STRING
-import com.vaticle.typedb.studio.test.integration.Utils.DOUBLE_CHEVRON_UP_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.DOUBLE_CHEVRON_DOWN_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.DOUBLE_CHEVRON_UP_ICON_STRING
+import com.vaticle.typedb.studio.test.integration.common.Data.SCHEMA_FILE_NAME
+import com.vaticle.typedb.studio.test.integration.common.Data.TQL_DATA_PATH
+import com.vaticle.typedb.studio.test.integration.common.Delays.NETWORK_IO
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.cloneAndOpenProject
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.connectToTypeDB
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.createDatabase
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.delayAndRecompose
+import com.vaticle.typedb.studio.test.integration.common.StudioActions.writeSchemaInteractively
+import com.vaticle.typedb.studio.test.integration.common.StudioTestHelpers.studioTestWithRunner
 import org.junit.Ignore
 import org.junit.Test
 
@@ -67,7 +68,7 @@ class TypeBrowserTest: IntegrationTest() {
             writeSchemaInteractively(composeRule, dbName = testID, SCHEMA_FILE_NAME)
 
             StudioState.client.session.tryOpen(database = testID, TypeDBSession.Type.DATA)
-            delayAndRecompose(composeRule, Delays.NETWORK_IO)
+            delayAndRecompose(composeRule, NETWORK_IO)
 
             composeRule.onAllNodesWithText("Project").get(0).performClick()
             composeRule.onAllNodesWithText("Project").get(1).performClick()
@@ -113,7 +114,7 @@ class TypeBrowserTest: IntegrationTest() {
             writeSchemaInteractively(composeRule, dbName = testID, SCHEMA_FILE_NAME)
 
             StudioState.client.session.tryOpen(database = testID, TypeDBSession.Type.DATA)
-            delayAndRecompose(composeRule, Delays.NETWORK_IO)
+            delayAndRecompose(composeRule, NETWORK_IO)
 
             StudioState.schema.exportTypeSchema { schema ->
                 StudioState.project.current!!.reloadEntries()
